refactor: move tests to integration folder
